requirements.txt
streamlit
streamlit-option-menu
streamlit_toggle
pandas
plotly
scipy
statsmodels
python-dotenv
qrcode
